simplify integration testing
